Replaced getInitialProps with getServerSideProps
