+++ examples/with-cookie-auth/utils/auth.js
@@ -1,4 +1,3 @@
-import { useEffect } from 'react'
 import Router from 'next/router'
 import nextCookie from 'next-cookies'
 import cookie from 'js-cookie'
@@ -29,38 +28,4 @@ export const logout = () => {
   // to support logging out from all windows
   window.localStorage.setItem('logout', Date.now())
   Router.push('/login')
-}
-
-export const withAuthSync = WrappedComponent => {
-  const Wrapper = props => {
-    const syncLogout = event => {
-      if (event.key === 'logout') {
-        console.log('logged out from storage!')
-        Router.push('/login')
-      }
-    }
-
-    useEffect(() => {
-      window.addEventListener('storage', syncLogout)
-
-      return () => {
-        window.removeEventListener('storage', syncLogout)
-        window.localStorage.removeItem('logout')
-      }
-    }, [])
-
-    return <WrappedComponent {...props} />
-  }
-
-  Wrapper.getInitialProps = async ctx => {
-    const token = auth(ctx)
-
-    const componentProps =
-      WrappedComponent.getInitialProps &&
-      (await WrappedComponent.getInitialProps(ctx))
-
-    return { ...componentProps, token }
-  }
-
-  return Wrapper
 }
